+++ pkg/providers/datastream/stream_test.go
@@ -7,6 +7,7 @@ import (
 	"github.com/tj/assert"
 
 	"github.com/akamai/AkamaiOPEN-edgegrid-golang/v8/pkg/datastream"
+	"github.com/akamai/terraform-provider-akamai/v6/pkg/common/tf"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )
 
@@ -198,7 +199,7 @@ func TestInterfaceSliceToStringSlice(t *testing.T) {
 
 	for name, test := range tests {
 		t.Run(name, func(t *testing.T) {
-			assert.Equal(t, test.expected, InterfaceSliceToStringSlice(test.input))
+			assert.Equal(t, test.expected, tf.InterfaceSliceToStringSlice(test.input))
 		})
 	}
 }