+++ pkg/providers/iam/testdata/TestResourceUserLifecycle/create_basic_notification_enable_email_notifications_field_missing.tf
@@ -9,6 +9,19 @@ resource "akamai_iam_user" "test" {
   country          = "country"
   phone            = "(111) 111-1111"
   enable_tfa       = false
-  enable_mfa       = false
+  enable_mfa       = true
   auth_grants_json = "[{\"groupId\":0,\"groupName\":\"group\",\"roleDescription\":\"\",\"roleName\":\"\"}]"
+  user_notifications {
+    options {
+      api_client_credential_expiry_notification = true
+      new_user_notification                     = true
+      password_expiry                           = true
+      proactive = [
+        "EdgeScape",
+      ]
+      upgrade = [
+        "NetStorage",
+      ]
+    }
+  }
 }