+++ pkg/providers/iam/testdata/TestResourceUserLifecycle/create_basic_notification_options_block_missing.tf
@@ -5,10 +5,13 @@ provider "akamai" {
 resource "akamai_iam_user" "test" {
   first_name       = "John"
   last_name        = "Smith"
-  email            = "jsmith@update.com"
+  email            = "jsmith@example.com"
   country          = "country"
   phone            = "(111) 111-1111"
   enable_tfa       = false
-  password         = "Password@123"
+  enable_mfa       = true
   auth_grants_json = "[{\"groupId\":0,\"groupName\":\"group\",\"roleDescription\":\"\",\"roleName\":\"\"}]"
+  user_notifications {
+    enable_email_notifications = true
+  }
 }