+++ pkg/providers/iam/testdata/TestResourceUserLifecycle/create_basic_with_password.tf
@@ -5,10 +5,11 @@ provider "akamai" {
 resource "akamai_iam_user" "test" {
   first_name       = "John"
   last_name        = "Smith"
-  email            = "jsmith@update.com"
+  email            = "jsmith@example.com"
   country          = "country"
   phone            = "(111) 111-1111"
   enable_tfa       = false
+  enable_mfa       = false
   password         = "Password@123"
   auth_grants_json = "[{\"groupId\":0,\"groupName\":\"group\",\"roleDescription\":\"\",\"roleName\":\"\"}]"
 }